pyproject.toml
[project]
name = "capture"
version = "0.1.0"
description = "Auto capture PowerPoint slides from browser and export to PPTX"
requires-python = ">=3.11"
dependencies = [
    "click>=8.3.1",
    "mss>=10.1.0",
    "numpy>=2.4.2",
    "pillow>=12.1.1",
    "playwright>=1.58.0",
    "pyautogui>=0.9.54",
    "pynput>=1.8.1",
    "python-pptx>=1.0.2",
]

[project.scripts]
capture = "main:cli"
